Vents now default to fully open, LED blinks when not connected to wifi, fixed LED toggle
When vents are powered on, they now go to fully open by default.

The LED blinks on/off every 200 ms when connecting to wifi.  If the connection is lost, it blinks continuously (100 ms / 300ms) until the conection is restablished.

Fixed the LED toggle so that it is properly set after using Blink LED and other things that change the LED state.