Auto-reboot if disconnected for 5 minutes, defines for open/closed state
If wifi is lost for >5 minutes, the device now reboots.  Sometimes it would get stuck in this state, and power-cycling was the only way to fix it, and having the software do that itself is cleaner.  It would be nice if it would just connect, but this is fine.  The vent is opened fully before rebooting.

Also added defines for fully opened and clsoed states to make the code more readable.